# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/forgot-password/confirm": {
      "post": {
        "operationId": "AuthController_confirmForgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/change-password": {
      "patch": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/farms": {
      "post": {
        "operationId": "FarmController_createFarm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFarmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Farms"
        ]
      },
      "get": {
        "operationId": "FarmController_findFarms",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Farms"
        ]
      }
    },
    "/farms/{id}": {
      "get": {
        "operationId": "FarmController_findFarm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Farms"
        ]
      },
      "patch": {
        "operationId": "FarmController_updateFarm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFarmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Farms"
        ]
      },
      "delete": {
        "operationId": "FarmController_removeFarm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Farms"
        ]
      }
    },
    "/farmers": {
      "get": {
        "operationId": "FarmController_findFarmers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Farmers"
        ]
      }
    },
    "/farmers/me": {
      "get": {
        "operationId": "FarmController_findMyFarmer",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Farmers"
        ]
      }
    },
    "/farmers/{id}": {
      "get": {
        "operationId": "FarmController_findFarmer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Farmers"
        ]
      }
    },
...
  },
  "info": {
    "title": "laicos API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "ada@example.com"
          },
          "password": {
            "type": "string",
            "example": "Password1!"
          },
          "firstName": {
            "type": "string",
            "example": "Ada",
            "description": "User first name"
          },
          "lastName": {
            "type": "string",
            "example": "Okafor",
            "description": "User last name"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+2348012345678",
            "description": "Phone number in E.164 format"
          },
          "role": {
            "enum": [
              "FARMER",
              "BUYER",
              "EXTENSION_AGENT",
              "RIDER"
            ],
            "type": "string",
            "example": "FARMER",
            "description": "Account role. Only FARMER or BUYER may self-register."
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName",
          "phoneNumber",
          "role"
        ]
      },
      "Role": {
        "type": "string",
        "enum": [
          "FARMER",
          "BUYER",
          "EXTENSION_AGENT",
          "RIDER",
          "ADMIN"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firebaseUid": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "example": "Ada"
          },
          "lastName": {
            "type": "string",
            "example": "Okafor"
          },
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+2348012345678"
          },
          "role": {
            "example": "FARMER",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "agreedToTerms": {
            "type": "boolean"
          },
          "isVerified": {
            "type": "boolean",
            "description": "Mirrors Firebase's emailVerified"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "firebaseUid",
          "firstName",
          "lastName",
          "email",
          "phoneNumber",
          "role",
          "agreedToTerms",
          "isVerified",
          "createdAt",
          "updatedAt"
        ]
      },
      "AuthSessionDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "AMf-vXy1__fake-refresh-token"
          },
          "expiresIn": {
            "type": "string",
            "example": "3600",
            "description": "Access token lifetime in seconds, as a string"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "user"
        ]
      },
      "AuthSessionResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "$ref": "#/components/schemas/AuthSessionDto"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "ada@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "example": "Password1!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Firebase refresh token"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "AMf-vXy1__fake-refresh-token"
          },
          "expiresIn": {
            "type": "string",
            "example": "3600",
            "description": "Access token lifetime in seconds, as a string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn"
        ]
      },
      "TokenPairResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "$ref": "#/components/schemas/TokenPairDto"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "ada@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "nullable": true,
            "example": null
          },
          "message": {
            "type": "string",
            "example": "OK"
          }
        },
        "required": [
          "data",
          "message"
        ]
      },
      "ConfirmForgotPasswordDto": {
        "type": "object",
        "properties": {
          "oobCode": {
            "type": "string",
            "description": "The oobCode query param from the password reset email link"
          },
          "newPassword": {
            "type": "string",
            "minLength": 1,
            "example": "NewPassword1!"
          }
        },
        "required": [
          "oobCode",
          "newPassword"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "minLength": 1,
            "example": "OldPassword1!"
          },
          "newPassword": {
            "type": "string",
            "minLength": 1,
            "example": "NewPassword1!"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "CreateFarmDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "example": "Green Acres"
          },
          "location": {
            "type": "string",
            "maxLength": 255,
            "example": "Ogbomosho, Oyo State"
          },
          "size": {
            "type": "number",
            "minimum": 1,
            "example": 2.5,
            "description": "Size in hectares"
          }
        },
        "required": [
          "name",
          "location",
          "size"
        ]
      },
      "FarmDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "description": "Farmer.id of the owner"
          },
          "name": {
            "type": "string",
            "example": "Green Acres"
          },
          "location": {
            "type": "string",
            "example": "Ogbomosho, Oyo State"
          },
          "size": {
            "type": "number",
            "example": 2.5,
            "description": "Size in hectares"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ownerId",
          "name",
          "location",
          "size",
          "createdAt",
          "updatedAt"
        ]
      },
      "FarmResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "$ref": "#/components/schemas/FarmDto"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "perPage": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 42
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "perPage",
          "total",
          "totalPages"
        ]
      },
      "FarmListResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FarmDto"
            }
          },
          "metaData": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "message",
          "data",
          "metaData"
        ]
      },
      "UpdateFarmDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "example": "Green Acres"
          },
          "location": {
            "type": "string",
            "maxLength": 255,
            "example": "Ogbomosho, Oyo State"
          },
          "size": {
            "type": "number",
            "minimum": 1,
            "example": 2.5,
            "description": "Size in hectares"
          }
        }
      },
      "FarmerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "farmerId": {
            "type": "string",
            "example": "FRM-000123",
            "description": "Public farmer code"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "farmerId",
          "userId",
          "createdAt",
          "updatedAt"
        ]
      },
      "FarmerListResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FarmerDto"
            }
          },
          "metaData": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "message",
          "data",
          "metaData"
        ]
      },
      "FarmerResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "$ref": "#/components/schemas/FarmerDto"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
...
    }
  }
}